changed double type to float

+++ GameParticles - student/GameParticles/Math/Vect4D.cpp
@@ -6,13 +6,13 @@
 
 Vect4D::Vect4D()
 {
-	this->x = 0.0;
-	this->y = 0.0;
-	this->z = 0.0;
-	this->w = 1.0;
+	this->x = 0.0f;
+	this->y = 0.0f;
+	this->z = 0.0f;
+	this->w = 1.0f;
 }
 
-Vect4D::Vect4D( double tx, double ty, double tz, double tw )
+Vect4D::Vect4D(float tx, float ty, float tz, float tw)
 {
 	this->x = tx;
 	this->y = ty;
@@ -27,14 +27,14 @@ Vect4D::~Vect4D()
 
 void Vect4D::norm(Vect4D& out)
 {
-	double mag = sqrt( this->x * this->x + this->y * this->y + this->z * this->z );
+	float mag = sqrtf(this->x * this->x + this->y * this->y + this->z * this->z);
 
 	if( 0.0f < mag )
 	{
 		out.x = this->x / mag;
 		out.y = this->y / mag;
 		out.z = this->z / mag;
-		out.w = 1.0;
+		out.w = 1.0f;
 	}
 }
 
@@ -60,7 +60,7 @@ Vect4D Vect4D::operator - ( Vect4D t )
 	return out;
 }
 
-Vect4D Vect4D::operator *( double scale)
+Vect4D Vect4D::operator *(float scale)
 {
 	Vect4D tmp;
 
@@ -71,7 +71,7 @@ Vect4D Vect4D::operator *( double scale)
 	return tmp;
 }
 
-double& Vect4D::operator[]( VECT_ENUM e )
+float& Vect4D::operator[](VECT_ENUM e)
 {
 	switch(e)
 	{
@@ -101,7 +101,7 @@ void Vect4D::Cross( Vect4D& vin, Vect4D& vout)
 	vout.w = 1.0f;
 }
 
-void Vect4D::set(double tx, double ty, double tz, double tw)
+void Vect4D::set(float tx, float ty, float tz, float tw)
 {
 	this->x = tx;
 	this->y = ty;
