+++ GameParticles - student/GameParticles/Math/Vect4D.h
@@ -17,26 +17,26 @@ public:
 	friend class Matrix;
 
 	Vect4D();	
-	Vect4D( double tx, double ty, double tz, double tw = 1.0f );
+	Vect4D(float tx, float ty, float tz, float tw = 1.0f);
 	~Vect4D();
 
 	void norm( Vect4D &out );
-	void set( double tx, double ty, double tz, double tw = 1.0f);
+	void set(float tx, float ty, float tz, float tw = 1.0f);
 	
 	Vect4D operator + ( Vect4D t );
 	Vect4D operator - ( Vect4D t );
-	Vect4D operator * ( double scale );
+	Vect4D operator * (float scale);
 
 	void Cross(Vect4D &vin, Vect4D &vout);
 
-	double &operator[]( VECT_ENUM e);
+	float &operator[](VECT_ENUM e);
 
 
 private:
-	double x;
-	double y;
-	double z;
-	double w;
+	float x;
+	float y;
+	float z;
+	float w;
 };
 
 #endif  //Vect4D.h
