+++ GameParticles - student/GameParticles/ParticleEmitter.h
@@ -27,18 +27,18 @@ private:
 	Vect4D	start_position;
 	Vect4D	start_velocity;
 	
-	double	max_life;
+	float	max_life;
 	int		max_particles;
-	double	spawn_frequency;
-	double	last_spawn;
-	double	last_loop;
+	float	spawn_frequency;
+	float	last_spawn;
+	float	last_loop;
 	int		last_active_particle;
 
 	// added for speed efficiency
 	int bufferCount;
 	Vect4D	vel_variance;
 	Vect4D	pos_variance;
-	double	scale_variance;
+	float	scale_variance;
 	std::list<Particle> drawBuffer;
 	Particle *headParticle;
 	std::list<Particle> particle_list;
